Add link to Lancaster University guide
The Lancaster University guide to the micro:bit runtime is itself a valuable collection of documentation for the more advanced micro:bit user. Although not strictly a teaching resource, it is resource that teachers and users may both want to use and doesn't fit directly in any of the other categories. I've included it here as a single link to their documentation.